# Use custom marshaler for trie.Node