to minimize allocations

before

```
omer@omer-ThinkPad-E14-Gen-3:~/Documents/juno$ go test ./blockchain/ -bench=MainnetBlock -benchmem -memprofile mem.out
goos: linux
goarch: amd64
pkg: github.com/NethermindEth/juno/blockchain
cpu: AMD Ryzen 7 5700U with Radeon Graphics         
BenchmarkMainnetBlock-16              10         308859680 ns/op         9854500 B/op     265245 allocs/op
PASS
ok      github.com/NethermindEth/juno/blockchain        4.577s
```

after

```
omer@omer-ThinkPad-E14-Gen-3:~/Documents/juno$ go test ./blockchain/ -bench=MainnetBlock -benchmem -memprofile mem.out
goos: linux
goarch: amd64
pkg: github.com/NethermindEth/juno/blockchain
cpu: AMD Ryzen 7 5700U with Radeon Graphics         
BenchmarkMainnetBlock-16              10         268575622 ns/op         5724414 B/op     161778 allocs/op
PASS
ok      github.com/NethermindEth/juno/blockchain        4.164s
```